+++ crates/gridthorn/src/runtime/test/mod.rs
@@ -2,6 +2,7 @@ use crate::prelude::*;
 
 mod command;
 mod input;
+mod presentation;
 mod time;
 
 #[derive(Default)]